+++ agb/examples/animated_background.rs
@@ -38,7 +38,7 @@ fn main(mut gba: agb::Gba) -> ! {
     }
 
     bg.commit(&mut vram);
-    bg.show();
+    bg.set_visible(true);
 
     let mut i = 0;
     loop {